Refactoring the model round 2
