+++ common/src/main/java/org/apache/karaf/vineyard/common/API.java
@@ -16,6 +16,9 @@
  */
 package org.apache.karaf.vineyard.common;
 
+import java.util.Collection;
+import java.util.Map;
+
 /**
  * This is a API abstract view, containing attributes common to any concrete API in Vineyard.
  */
@@ -30,8 +33,18 @@ public abstract class API {
     /** Open text describing the API */
     private String description;
 
-    /** The API maintener */
-    private Maintainer maintainer;
+    /** The resource operation data format (json, xml, ...) */
+    private DataFormat format;
+
+    /** The policies define for the API */
+    private Collection<Policy> policies;
+
+    /** Additional data functional and technical related to the API */
+    private Map<String, String> metadata;
+
+    /** The resource endpoint location */
+    private String endpoint;
+
 
     public String getId() {
         return id;
@@ -57,11 +70,35 @@ public abstract class API {
         this.description = description;
     }
 
-    public Maintainer getMaintainer() {
-        return maintainer;
+    public DataFormat getFormat() {
+        return format;
     }
 
-    public void setMaintainer(Maintainer maintainer) {
-        this.maintainer = maintainer;
+    public void setFormat(DataFormat format) {
+        this.format = format;
+    }
+
+    public String getEndpoint() {
+        return endpoint;
+    }
+
+    public void setEndpoint(String endpoint) {
+        this.endpoint = endpoint;
+    }
+
+    public Collection<Policy> getPolicies() {
+        return policies;
+    }
+
+    public void setPolicies(Collection<Policy> policies) {
+        this.policies = policies;
+    }
+
+    public Map<String, String> getMetadata() {
+        return metadata;
+    }
+
+    public void setMetadata(Map<String, String> metadata) {
+        this.metadata = metadata;
     }
 }
