+++ common/src/main/java/org/apache/karaf/vineyard/common/RestAPIResource.java
@@ -16,9 +16,6 @@
  */
 package org.apache.karaf.vineyard.common;
 
-import java.util.Collection;
-import java.util.Map;
-
 /**
  * Describe a resource in a Rest API.
  */
@@ -29,20 +26,6 @@ public class RestAPIResource {
 
     /** The resource operation method */
     private String method;
-
-    /** The resource operation data format (json, xml, ...) */
-    private DataFormat format;
-
-    private Environment environment;
-
-    private long throttling;
-
-    private Collection<Policy> policies;
-
-    private Map<String, String> metadata;
-
-    /** The resource endpoint location */
-    private String endpoint;
 
     public String getPath() {
         return path;
@@ -60,51 +43,4 @@ public class RestAPIResource {
         this.method = method;
     }
 
-    public DataFormat getFormat() {
-        return format;
-    }
-
-    public void setFormat(DataFormat format) {
-        this.format = format;
-    }
-
-    public String getEndpoint() {
-        return endpoint;
-    }
-
-    public void setEndpoint(String endpoint) {
-        this.endpoint = endpoint;
-    }
-
-    public long getThrottling() {
-        return throttling;
-    }
-
-    public void setThrottling(long throttling) {
-        this.throttling = throttling;
-    }
-
-    public Collection<Policy> getPolicies() {
-        return policies;
-    }
-
-    public void setPolicies(Collection<Policy> policies) {
-        this.policies = policies;
-    }
-
-    public Environment getEnvironment() {
-        return environment;
-    }
-
-    public void setEnvironment(Environment environment) {
-        this.environment = environment;
-    }
-
-    public Map<String, String> getMetadata() {
-        return metadata;
-    }
-
-    public void setMetadata(Map<String, String> metadata) {
-        this.metadata = metadata;
-    }
 }
